move the initialize_block out to a new file
