+++ test/unit/flash_attention/flash_attention_prefill/flash_prefill_testbed_3x.hpp
@@ -44,7 +44,7 @@
 #include "flash_attention_v2/collective/xe_flash_attn_prefill_softmax_epilogue.hpp"
 #include "cutlass/util/GPU_Clock.hpp"
 #include "cutlass/util/sycl_event_manager.hpp"
-#include "cutlass/util/mixed_dtype_utils.hpp"
+#include "cutlass/util/initialize_block.hpp"
 
 #include <cute/tensor.hpp>
 #include <random>
